feat: convert resources manager dialog into page
